// devops-frontend/src/generated-sources/openapi/models/ModelPlugin.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NEW Devops API
 * Devops API Sec
 *
 * The version of the OpenAPI document: v0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ModelPlugin
 */
export interface ModelPlugin {
    /**
     * 
     * @type {boolean}
     * @memberof ModelPlugin
     */
    isDefault: boolean;
    /**
     * 
     * @type {string}
     * @memberof ModelPlugin
     */
    name: string;
}

/**
 * Check if a given object implements the ModelPlugin interface.
 */
export function instanceOfModelPlugin(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "isDefault" in value;
    isInstance = isInstance && "name" in value;

    return isInstance;
}

export function ModelPluginFromJSON(json: any): ModelPlugin {
    return ModelPluginFromJSONTyped(json, false);
}

export function ModelPluginFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelPlugin {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'isDefault': json['isDefault'],
        'name': json['name'],
    };
}

export function ModelPluginToJSON(value?: ModelPlugin | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'isDefault': value.isDefault,
        'name': value.name,
    };
}